requirements.txt
requests>=2.31.0
Pillow>=10.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
